+++ src/modules/nft/service_layer/nft.service.ts
@@ -163,13 +163,16 @@ export class NftService {
     const downsizedFile = await this.fileProcessingService.downsizeFile(file.path, file.mimetype);
     const uniqueFiles = [optimisedFile, downsizedFile].filter((fileItem) => fileItem.path !== file.path);
 
+    this.logger.log('Starting to upload files to AWS');
     await Promise.all([
       this.s3Service.uploadDocument(file.path, `${file.filename}`),
       ...uniqueFiles.map((fileItem) => this.s3Service.uploadDocument(fileItem.path, `${fileItem.fullFilename()}`)),
     ]);
-    const data = await fs.readFile(file.path);  
+    this.logger.log('Uploaded files to bucket successfully');
+    const data = await fs.readFile(file.path);
+    this.logger.log('Starting to store data to aweave');
     const arweaveUrl = await this.arweaveService.storeData(data, file.mimetype);
-
+    this.logger.log('Stored data to arweave successfully');
     await Promise.all(
       [file.path, ...uniqueFiles.map((file) => file.path)].map((path) => this.fileSystemService.removeFile(path)),
     );
@@ -241,8 +244,9 @@ export class NftService {
     }
     const bodyClass = plainToClass(GetNftTokenUriBody, { ...body });
     await this.validateReqBody(bodyClass);
-
+    this.logger.log('Starting to process file');
     const { optimisedFile, downsizedFile, arweaveUrl } = await this.processUploadedFile(file);
+    this.logger.log('Processed file successfully');
     const idxs = [...Array(bodyClass.numberOfEditions).keys()];
     const tokenUri = await this.generateTokenUri({
       name: bodyClass.name,
@@ -291,9 +295,14 @@ export class NftService {
     let mintingCollection = this.mintingCollectionRepository.create({
       userId,
       name: bodyClass.name,
-      description: bodyClass.description,
       symbol: bodyClass.symbol,
       coverUrl: s3Result.url,
+      description: bodyClass.description,
+      siteLink: bodyClass.siteLink,
+      discordLink: bodyClass.discordLink,
+      instagramLink: bodyClass.instagramLink,
+      mediumLink: bodyClass.mediumLink,
+      telegramLink: bodyClass.telegramLink,
     });
     mintingCollection = await this.mintingCollectionRepository.save(mintingCollection);
 
@@ -372,9 +381,8 @@ export class NftService {
       throw new NftCollectionBadOwnerException();
     }
 
-    const filteredAttributes = this.filterObjectAttributes(data, ['description']);
-    for (const attribute in filteredAttributes) {
-      collection[attribute] = filteredAttributes[attribute];
+    for (const attribute in data) {
+      collection[attribute] = data[attribute];
     }
 
     await this.nftCollectionRepository.save(collection);
@@ -395,14 +403,19 @@ export class NftService {
     }
   }
 
-  public async getSavedNfts(userId: number) {
-    const savedNfts = await this.savedNftRepository.find({
-      where: {
-        userId,
-      },
-    });
+  public async getSavedNfts(userId: number, limit = 8, offset = 0) {
+    const [savedNfts, count] = await this.savedNftRepository
+      .createQueryBuilder('savedNft')
+      .where('savedNft.userId = :userId', { userId })
+      .leftJoinAndMapOne('savedNft.collection', NftCollection, 'collection', 'collection.id = savedNft.collectionId')
+      .take(limit)
+      .limit(offset)
+      .getManyAndCount();
 
-    return savedNfts;
+    return {
+      nfts: savedNfts,
+      pagination: this.paginationDetails(count, limit, offset),
+    };
   }
 
   public async editSavedNft(id: number, userId: number, params: EditSavedNftParams) {
@@ -490,13 +503,16 @@ export class NftService {
   };
 
   private async reduceUserNftsByEdition(
-    userId: number,
+    userAddress: string,
     additionalData?: NftAdditionalData,
     prefetchData?: NftPrefetchData,
+    limit = 8,
+    offset = 0,
+    name = '',
+    collectionsQuery = '',
   ) {
     let nfts = [];
     const query = this.nftRepository.createQueryBuilder('nft');
-    const user = await this.userRepository.findOne({ where: { id: userId } });
 
     if (additionalData?.creator && !prefetchData?.creator) {
       query.leftJoinAndMapOne('nft.creator', User, 'creator', 'creator.address = nft.creator');
@@ -506,7 +522,41 @@ export class NftService {
       query.leftJoinAndMapOne('nft.owner', User, 'owner', 'owner.address = nft.owner');
     }
 
-    nfts = await query.where('nft.owner = :owner', { owner: user.address }).orderBy('nft.createdAt', 'DESC').getMany();
+    const editionsCount = parseInt(
+      (
+        await this.nftRepository.query(
+          'SELECT COUNT(DISTINCT "editionUUID") FROM "universe-backend"."nft" WHERE "nft"."owner" = $1',
+          [userAddress],
+        )
+      )[0].count,
+    );
+
+    let conditions =
+      'nft.editionUUID IN (' +
+      'SELECT "editionUUID" FROM (' +
+      'SELECT DISTINCT "editionUUID", MAX("id")' +
+      'FROM "universe-backend"."nft" as "nft" ' +
+      'WHERE "nft"."owner" = :owner ';
+
+    if (name) {
+      conditions += 'AND (LOWER(nft.name) LIKE :name)';
+    }
+
+    if (collectionsQuery) {
+      conditions += 'AND nft.collectionId IN(:...collections)';
+    }
+
+    conditions += 'GROUP BY "editionUUID"' + ') AS "nft" ORDER BY "max" DESC LIMIT :limit OFFSET :offset)';
+
+    query.andWhere(conditions, {
+      owner: userAddress,
+      name: `${name.toLowerCase()}%`,
+      collections: collectionsQuery.split(',').map((id) => Number(id)),
+      limit: limit,
+      offset: offset,
+    });
+
+    nfts = await query.orderBy('nft.id', 'DESC').getMany();
 
     if (prefetchData?.owner) {
       nfts = nfts.map((nft) => {
@@ -546,6 +596,7 @@ export class NftService {
           tokenIds,
         };
       }),
+      pagination: this.paginationDetails(editionsCount, limit, offset),
     };
   }
 
@@ -583,17 +634,17 @@ export class NftService {
     };
   }
 
-  public async getMyNfts(userId: number) {
+  public async getMyNfts(userAddress: string, limit = 8, offset = 0) {
     const additionalData: NftAdditionalData = {
       collection: false,
       owner: false,
       creator: true,
     };
 
-    return await this.reduceUserNftsByEdition(userId, additionalData);
+    return await this.reduceUserNftsByEdition(userAddress, additionalData, null, limit, offset);
   }
 
-  public async getUserNfts(username: string) {
+  public async getUserNfts(username: string, limit = 8, offset = 0, name = '', collections = '') {
     const user = await this.userRepository.findOne({ where: [{ universePageUrl: username }, { address: username }] });
 
     if (!user) {
@@ -610,7 +661,7 @@ export class NftService {
       owner: user,
     };
 
-    return this.reduceUserNftsByEdition(user.id, additionalData, prefetchData);
+    return this.reduceUserNftsByEdition(user.address, additionalData, prefetchData, limit, offset, name, collections);
   }
 
   public async getMyNftsAvailability(userId: number, start = 0, limit = 8, size = 0, auctionId = 0) {
@@ -676,28 +727,58 @@ export class NftService {
 
     return {
       nfts: mappedNfts,
-      pagination: {
-        page: start === 0 ? 1 : Math.ceil(start / limit + 1),
-        hasNextPage: editionsCount > +start + +limit,
-        totalPages: Math.ceil(editionsCount / limit),
-      },
+      pagination: this.paginationDetails(editionsCount, limit, start),
     };
   }
 
-  public async getMyNftsCollections(userId: number) {
-    const user = await this.userRepository.findOne({ where: { id: userId } });
-
+  public async getMyNftsCollections(address: string) {
     const nfts = await this.nftRepository.find({
-      where: { owner: user.address },
-      select: ['collectionId'],
+      where: { owner: address },
+      select: ['collectionId', 'editionUUID'],
     });
 
     const collections = await this.nftCollectionRepository.find({
       where: { id: In(nfts.map((nft) => nft.collectionId)) },
     });
 
+    const mappedCollections = collections.map((collection) => {
+      const editionsReducedNfts = nfts.reduce((acc, nft) => {
+        if (!acc[nft.collectionId]) [(acc[nft.collectionId] = [])];
+
+        const hasNFt = acc[nft.collectionId].find((nftColl) => nftColl.editionUUID === nft.editionUUID);
+        if (!hasNFt) {
+          acc[nft.collectionId].push(nft);
+        }
+
+        return acc;
+      }, {});
+
+      return {
+        ...classToPlain(collection),
+        nftCount: editionsReducedNfts[collection.id].length,
+      };
+    });
+
+    return {
+      collections: mappedCollections,
+    };
+  }
+
+  public async getMyCollectionsTab(userId: number, limit = 8, offset = 0) {
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+
+    const [collections, count] = await this.nftCollectionRepository
+      .createQueryBuilder('collection')
+      .where('collection.owner = :userAddress', { userAddress: user.address })
+      .orWhere('collection.publicCollection = TRUE')
+      .take(limit)
+      .skip(offset)
+      .orderBy({ id: 'DESC' })
+      .getManyAndCount();
+
     return {
       collections: collections.map((collection) => classToPlain(collection)),
+      pagination: this.paginationDetails(count, limit, offset),
     };
   }
 
@@ -744,23 +825,25 @@ export class NftService {
       )[0].count,
     );
 
-    const conditions =
+    let conditions =
       'nft.editionUUID IN (' +
-      'SELECT DISTINCT("nft"."editionUUID") FROM (' +
-      'SELECT "editionUUID", "id" FROM "universe-backend"."nft" as "nft" WHERE "nft"."collectionId" = :collectionId ORDER BY "nft"."id" DESC' +
-      ') AS "nft" LIMIT :limit OFFSET :offset' +
-      ')';
+      'SELECT "editionUUID" FROM (' +
+      'SELECT DISTINCT "editionUUID", MAX("id")' +
+      'FROM "universe-backend"."nft" as "nft" ' +
+      'WHERE "nft"."collectionId" = :collectionId ';
+
+    if (name) {
+      conditions += 'AND (LOWER(nft.name) LIKE :name)';
+    }
+
+    conditions += 'GROUP BY "editionUUID"' + ') AS "nft" ORDER BY "max" DESC LIMIT :limit OFFSET :offset)';
 
     const query = this.nftRepository
       .createQueryBuilder('nft')
       .leftJoinAndMapOne('nft.owner', User, 'owner', 'owner.address = nft.owner')
       .leftJoinAndMapOne('nft.creator', User, 'creator', 'creator.address = nft.creator')
-      .where(conditions, { collectionId: collection.id, limit: limit, offset: offset })
+      .where(conditions, { collectionId: collection.id, limit: limit, offset: offset, name: `%${name}%` })
       .orderBy('nft.createdAt', 'DESC');
-
-    if (name) {
-      query.andWhere('LOWER("nft"."name") LIKE :name', { name: `%${name}%` });
-    }
 
     const nfts = await query.getMany();
 
@@ -805,30 +888,20 @@ export class NftService {
     return {
       collection: classToPlain(collection),
       nfts: formattedNfts,
-      pagination: {
-        totalCount: editionsCount,
-        page: Math.ceil(offset / limit + 1),
-        hasNextPage: editionsCount > offset + limit,
-        totalPages: Math.ceil(editionsCount / limit),
-      },
+      pagination: this.paginationDetails(editionsCount, limit, offset),
     };
   }
 
-  public async getMyNftsPage(userId: number) {
+  public async getMyNftsPage(userAddress: any, limit = 8, offset = 0, name = '', collections = '') {
     const additionaData: NftAdditionalData = {
       collection: false,
       owner: false,
       creator: true,
     };
-    const mintedNfts = await this.reduceUserNftsByEdition(userId, additionaData);
-
-    return {
-      nfts: mintedNfts.nfts,
-      pagination: {},
-    };
+    return await this.reduceUserNftsByEdition(userAddress, additionaData, null, limit, offset, name, collections);
   }
 
-  public async getMyNftsPendingPage(userId) {
+  public async getMyNftsPendingPage(userId, limit = 8, offset = 0) {
     const mintingNfts = await this.mintingNftRepository.find({
       where: { userId, txStatus: 'pending' },
       order: { createdAt: 'DESC' },
@@ -843,7 +916,6 @@ export class NftService {
         ...classToPlain(nft),
         collection: classToPlain(idCollectionMap[nft.collectionId]),
       })),
-      pagination: {},
     };
   }
 
@@ -897,8 +969,6 @@ export class NftService {
 
     return {
       collections: mintingCollections.map((mintingCollection) => classToPlain(mintingCollection)),
-      // TODO: Future object which will container pagination information
-      pagination: {},
     };
   }
 
@@ -911,6 +981,35 @@ export class NftService {
 
     return {
       count: mintingCollectionCount,
+    };
+  }
+
+  public async getMyNftsSummary(address: string) {
+    const user = await this.userRepository.findOne({ where: { address } });
+    if (!user) {
+      throw new UserNotFoundException();
+    }
+
+    const editionsCount = parseInt(
+      (
+        await this.nftRepository.query(
+          'SELECT COUNT(DISTINCT "editionUUID") FROM "universe-backend"."nft" WHERE "nft"."owner" = $1',
+          [user.address.toLowerCase()],
+        )
+      )[0].count,
+    );
+
+    const collectionsCount = await this.nftCollectionRepository.count({
+      where: [{ owner: user.address }, { publicCollection: true }],
+    });
+
+    const savedNftsCount = await this.savedNftRepository.count({
+      where: { userId: user.id },
+    });
+    return {
+      nfts: editionsCount,
+      collections: collectionsCount,
+      savedNfts: savedNftsCount,
     };
   }
 
@@ -931,5 +1030,18 @@ export class NftService {
     }
     const urlComponents = url.split(/[.]+/);
     return urlComponents[urlComponents.length - 1];
+  }
+
+  private paginationDetails(itemsCount: number, limit: number, offset: number) {
+    const page = Math.ceil(offset / limit + 1);
+    const hasNextPage = itemsCount > offset + limit;
+    const totalPages = Math.ceil(itemsCount / limit);
+
+    return {
+      totalCount: itemsCount,
+      page: page,
+      hasNextPage: hasNextPage,
+      totalPages: totalPages,
+    };
   }
 }